Apply filters before delivering any data
As we discussed drinking beers, we need to make a clean distinction between
precision and recall.

To have a complete result ser a searching by Criteria should be used to bring
large amount of peers, with evasive criteria as possible

To achieve more precision on the requested datas, one should give to peers a list
of precise metadatas he is waiting for.

It appear to be easier for a peer to limit the number of output than for the
client to receive a large amount and filter it after.

For example, if a user want to retrieve every audio file on the network matching
a Metal music style and a byte rate of 320B/s, it will be more comfy to make on
the first place a search on the DHT on :

Source   : music
Metadata : style:Metal

Added to it a requested metadata may be used to ask distant peers to only give
music having a metada byterate matching with 320.

Later, it will be discussed to add the possibility to use operands on those
metadata filter.