Add consistency check hook
As we now give the distant peer the ability to filter the returned data. We have
the possibility to check if the distant peer has done a proper work or not.

The main idea is to bring the opportunity to link a peer and which data he gave.
Make a client side check and alert the system if the work wasn't done properly.

For now juste basic mecanism is in da place.

MetaPluginAPI is now able to make operations call core hooks before any plugin
asynchronous call.